feat: kafka producer and consumer

+++ v1/src/repositories/user/user.go
@@ -5,11 +5,11 @@ import (
 )
 
 func (repo *UserRepositoryStruct) GetUserById(id float64) (r consts.UserResponse, err error) {
-	query := `SELECT id,email,name,status,created_at 
+	query := `SELECT id,email,name,status,created_at
 				FROM "users"
 				WHERE id = $1`
 
-	err = repo.DB.QueryRowx(query, id).StructScan(&r)
+	err = repo.DB.QueryRowx(query, int64(id)).StructScan(&r)
 
 	if err != nil {
 		return r, err
